Set up docker-compose development environment
For the static site conversion, we need an isolated python environment during development whereby python libraries can be easily installed and the project run. Setting up a pythonenvironment using the virtualenv tool exposed that local python setups can be variable and frustrating to set up, thereby it would be easier to work off a python image in docker containers.

Refs #1086