Add plugin_loader script
This is a direct port of the PluginLoader.cs helper file from C# to
Python. These functions are used to load, validate, and otherwise setup
the plugins for use in the framework.

In addition to the functions contained in PluginLoader.cs, the
GetPluginConfigurationData from Geosite.cs is ported here as well.
Adding this function here instead of main.py seems to be a better fit
with the purpose of this file.

Refs #1090